backend/requirements.txt
Flask>=2.3.3,<3.1
flask-cors>=3.0.10,<4.0.0
flask-sqlalchemy>=3.0.0
flask-session>=0.5.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
numpy>=1.26,<3
pandas>=2.0,<3
scikit-learn>=1.3,<1.7
joblib>=1.2,<2
requests>=2.28.0
Pillow>=10.0.0
transformers>=4.0.0
torch>=2.0.0
tf-keras>=2.15.0
sentencepiece>=0.1.99
easyocr>=1.7.0
PyPDF2>=3.0.0
groq>=0.4.0
pytesseract>=0.3.10
opencv-python>=4.8.0